+++ packages/docs/docs/main/change-log.md
@@ -24,6 +24,8 @@ This page lists the changes to the IsaacScript framework.
   - `getNewRoomCandidatesBesideRoom` (returns an array of objects instead of tuples)
   - `getNewRoomCandidatesForLevel` (returns an array of objects instead of tuples)
   - `getNewRoomCandidate` (returns an array of objects instead of tuples)
+  - `anyPlayerHasCollectible` (now takes an optional param of `ignoreModifiers`)
+  - `anyPlayerHasTrinket` (now takes an optional param of `ignoreModifiers`)
 - Added the following custom callbacks:
   - `POST_TEAR_KILL` (Thanks 4Grabs)
   - `POST_PROJECTILE_KILL` (Thanks 4Grabs)